Move the config modifier function
