+++ tardis-setups/setup_utils/__init__.py
@@ -1,0 +1,1 @@
+from setup_utils.config import *
